make fixes to column inference from testing

+++ rasgotransforms/rasgotransforms/transforms/target_encode/target_encode.py
@@ -1,3 +1,3 @@
 def infer_columns(args, source_columns) -> dict:
-    source_columns[f"{args['column']}_target_encoded"] = 'decimal'
+    source_columns[f"{args['column']}_target_encoded"] = 'float'
     return source_columns
